Attempt to fix accidental editing by double click
Prompts:
```
Really nice. The only problem is that i accidentally click on the labels when i intend to increase the counter. Could you solve that somehow?

This seems to work exactly like the previous version. What i meant is that the edit boxes for the labels are in the middle of the clickable are where i click to increase the counter. So i accidentally open the edit dialog when i'm not looking at my phone. I intend to use this app on my mobile to count cars on the road so i'm looking at the road and not my phone.
```